Separate code files between feature sets
